experiments/requirements.txt
nltk
networkx>=2
plotly
matplotlib
pandas 
numpy
pytorch